requirements.txt
django
djangorestframework
django-oauth-toolkit
MySQL-python
django-registration
